backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "proyect test"
authors = ["JhonAngulo <jhonmanuelm89@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pylint = "^3.2.5"
sqlalchemy = "^2.0.31"
databases = {extras = ["asyncpg"], version = "^0.9.0"}
pyjwt = "^2.8.0"
pydantic = "^2.8.2"
psycopg2-binary = "^2.9.9"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
